requirements.txt
opencv-python==4.8.1.78
-f https://download.pytorch.org/whl/torch_stable.html torch==1.11.0+cu113
-f https://download.pytorch.org/whl/torch_stable.html torchvision==0.12.0+cu113
segmentation-models-pytorch==0.2.0
albumentations==1.1.0
scikit-image==0.18.3
scikit-learn==1.3.2
scipy==1.11.4
Pillow==8.4.0
matplotlib==3.8.2
pandas==2.1.3